refactor: desktop 컴포넌트에서 calendar 부분 분리해서 관리
